// examples/workflow/main/src/lib/model/enum/actions/extended-user-action-params.enum.ts
/*
 * Copyright(c) RIB Software GmbH
 */

/**
 * Input/Output parameters for the action `Extended User` Action.
 */
export enum ExtendedUserActionParams {
	Html = 'HTML',
	Script = 'Script',
	Context = 'Context',
	IsPopupUp = 'IsPopUp',
	EvaluateProxy = 'EvaluateProxy',
	DisableRefresh = 'DisableRefresh',
	AllowReassign = 'AllowReassign',
	Title = 'Title',
	Subtitle = 'Subtitle',
	DialogConfig = 'DialogConfig'
}