// examples/workflow/main/src/lib/model/enum/actions/portal-review-bidder-invitation-email-action-params.enum.ts
/*
 * Copyright(c) RIB Software GmbH
 */

/**
 * Input/Output parameters for the action 'Portal Review Bidder Invitation Email'
 */
export enum PortalReviewBidderInvitationEmailActionEditorParams {
	Recipient = 'recipient',
	Subject = 'subject',
	Body = 'body',
	ResultCode = 'ResultCode',
	IsPopUp = 'IsPopUp',
}